requirements.txt
langchain-core
langchain-community
langchain
langchain-google-genai
langchain-groq
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
sentence-transformers
streamlit